improved test suite.

+++ src/test/java/com/joshlong/mogul/api/podcasts/production/AudioEncoderTest.java
@@ -1,26 +1,36 @@
 package com.joshlong.mogul.api.podcasts.production;
 
 import com.joshlong.mogul.api.media.AudioEncoder;
+import org.junit.jupiter.api.Test;
+import org.springframework.core.io.ClassPathResource;
 import org.springframework.util.Assert;
-import org.springframework.util.SystemPropertyUtils;
+import org.springframework.util.FileCopyUtils;
 
 import java.io.File;
+import java.io.FileOutputStream;
+import java.nio.file.Files;
 
 class AudioEncoderTest {
 
 	private final AudioEncoder encoder = new AudioEncoder();
 
-	AudioEncoderTest() throws Exception {
-	}
-
+	@Test
 	void transcodeWavToMp3s() throws Exception {
-		var wav = new File(
-				SystemPropertyUtils.resolvePlaceholders("${HOME}/Desktop/misc/sample-podcast/test.wav ".trim()));
-		Assert.state(wav.exists(), "the .wav file exists");
-		var output = encoder.encode(wav);
-		Assert.state(output.length() < wav.length(), "the new file should be a _lot_ smaller than the original!");
-		Assert.state(output.getName().endsWith(".mp3"), "the new file should be an .mp3");
-		output.delete();
+		var output = (File) null;
+		try {
+			var input = new ClassPathResource("/samples/input.wav");
+			var wav = Files.createTempFile("input", ".wav").toFile();
+			FileCopyUtils.copy(input.getInputStream(), new FileOutputStream(wav));
+			Assert.state(wav.exists(), "the .wav file exists");
+			output = encoder.encode(wav);
+			Assert.state(output.length() < wav.length(), "the new file should be a _lot_ smaller than the original!");
+			Assert.state(output.getName().endsWith(".mp3"), "the new file should be an .mp3");
+		}
+		finally {
+			if (null != output)
+				output.delete();
+		}
+
 	}
 
 }
